Divide configure-algorithm.tsx to components
