+++ web-app/client/src/components/AlgorithmFormConfigurator/FormHeader.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import styles from '@styles/ConfigureAlgorithm.module.scss';
+
+const FormHeader = (
+  <>
+    <h2 className={styles.name_main}>Configure Algorithm</h2>
+    <h6 className={styles.description}>Select algorithm parameters</h6>
+  </>
+);
+
+export default FormHeader;
