+++ web-app/client/src/components/AlgorithmFormConfigurator/index.ts
@@ -1,0 +1,3 @@
+import AlgorithmFormConfigurator from './AlgorithmFormConfigurator';
+
+export default AlgorithmFormConfigurator;
